finished project

+++ controllers/studentController.js
@@ -1,55 +1,52 @@
-const {Student, Course, StudentCourses} = require('../models')
+const {Student, Course, StudentCourses} = require('../models');
 
 //view all
 module.exports.viewAll = async function (req, res) {
     const students = await Student.findAll();
     res.render('student/view_all', {students});
-}
+};
 
 //profile
 module.exports.viewProfile = async function (req, res) {
-    const student = await Student.findByPk(req.params.id, {
-        include: 'courses'
-    });
-    const courses = await Course.findAll();
-    let availableCourses = [];
-    for (let i = 0; i < courses.length; i++) {
-        if (!studentHasCourse(student, courses[i])) {
-            availableCourses.push(courses[i]);
-        }
-    }
-    res.render('student/profile', {student, availableCourses})
-}
+    const isAdmin = req.user.can('view students');
+    const profileBelongsToUser = req.user.can('view self') && req.user.matchesUserId(req.params.id);
 
-//render add
-module.exports.renderAddForm = function (req, res) {
-    const student = {
-        first_name: '',
-        last_name: '',
-        grade_level: 9,
+    //if they arent staff and they arent the student whose profile they are trying to view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
     }
-    res.render('student/add', {student});
-}
-
-//add
-module.exports.addStudent = async function (req, res) {
-    const student = await Student.create({
-        first_name: req.body.first_name,
-        last_name: req.body.last_name,
-        grade_level: req.body.grade_level
-    });
-    res.redirect(`/students/profile/${student.id}`);
-}
+    const students = await Student.findAll();
+    res.render('students/view_all', {students});
+};
 
 //render edit
 module.exports.renderEditForm = async function (req, res) {
+    const isAdmin = req.user.can('edit student');
+    const profileBelongsToUser = req.user.can('edit self') && req.user.matchesUserId(req.params.id);
+
+    //if they arent staff and they arent the student whose profile they are trying to view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
+
     const student = await Student.findByPk(req.params.id);
     console.log(student);
     res.render('student/edit', {student});
-}
+};
 
 //update
 module.exports.updateStudent = async function (req, res) {
+    const isAdmin = req.user.can('edit student');
+    const profileBelongsToUser = req.user.can('edit self') && req.user.matchesUserId(req.params.id);
+
+    //if they arent staff and they arent the student whose profile they are trying to view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
+
     const student = await Student.update({
         first_name: req.body.first_name,
         last_name: req.body.last_name,
@@ -61,31 +58,50 @@ module.exports.updateStudent = async function (req, res) {
         }
     });
     res.redirect(`/students/profile/${req.params.id}`);
-}
+};
 
 //delete
 module.exports.deleteStudent = async function (req, res) {
+    if (!user.can('delete student')){
+        res.redirect('/');
+        return
+    }
     await Student.destroy({
         where: {
             id: req.params.id
         }
     });
     res.redirect('/students');
-}
+};
 
 //Add course to student
 module.exports.enrollStudent = async function (req, res) {
+    const isAdmin = req.user.can('enroll student');
+    const profileBelongsToUser = req.user.can('enroll self') && req.user.matchesUserId(req.params.id);
 
+    //if they arent staff and they arent the student whose profile they are trying to view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
         await StudentCourses.create({
             student_id: req.params.studentId,
             course_id: req.body.course
-        })
+        });
         res.redirect(`/students/profile/${req.params.studentId}`);
 
-}
+};
 
 //delete course from student
 module.exports.removeCourse = async function(req, res){
+    const isAdmin = req.user.can('drop student');
+    const profileBelongsToUser = req.user.can('drop self') && req.user.matchesUserId(req.params.id);
+
+    //if they arent staff and they arent the student whose profile they are trying to view, redirect
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/');
+        return
+    }
     await StudentCourses.destroy({
         where: {
             student_id: req.params.studentId,
@@ -93,7 +109,7 @@ module.exports.removeCourse = async function(req, res){
         }
     });
     res.redirect(`/students/profile/${req.params.studentId}`)
-}
+};
 
 function studentHasCourse(student, course) {
     for (let i = 0; i < student.courses.length; i++) {
